Refactor radiator API logic out of main and document code blocks
